+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Initial release of jira-download
+- Initial release of jarkdown
 - Export single Jira Cloud issues to Markdown format
 - Download and embed all attachments locally
 - Convert HTML descriptions to clean Markdown
@@ -44,5 +44,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Comments not included (planned for v0.2.0)
 - Jira Cloud only (Server/Data Center support planned)
 
-[Unreleased]: https://github.com/chrisbyboston/jira-download/compare/v0.1.0...HEAD
-[0.1.0]: https://github.com/chrisbyboston/jira-download/releases/tag/v0.1.0
+[Unreleased]: https://github.com/chrisbyboston/jarkdown/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/chrisbyboston/jarkdown/releases/tag/v0.1.0